requirements.txt
amqp==2.4.2
attrs==19.1.0
autopep8==1.4.3
beautifulsoup4==4.7.1
cachetools==3.1.0
Cerberus==1.2
certifi==2019.3.9
chardet==3.0.4
Click==7.0
cssselect==1.0.3
cssutils==1.0.2
dnspython==1.16.0
eventlet==0.24.1
flasgger==0.9.2
Flask==1.0.2
greenlet==0.4.15
grpcio==1.19.0
grpcio-tools==1.19.0
idna==2.8
importlib-metadata==0.8
itsdangerous==1.1.0
Jinja2==2.10
jsonschema==3.0.1
kombu==4.4.0
lxml==4.3.2
MarkupSafe==1.1.1
mistune==0.8.4
mock==2.0.0
monotonic==1.5
multi-key-dict==2.0.3
mysql-connector==2.1.6
nameko==2.11.0
path.py==11.5.0
pbr==5.1.3
pika==0.13.1
premailer==3.4.0
protobuf==3.7.0
pycodestyle==2.5.0
pynliner==0.8.0
pyrsistent==0.14.11
python-jenkins==1.4.0
PyYAML==3.13
requests==2.21.0
six==1.12.0
soupsieve==1.8
urllib3==1.24.1
vine==1.2.0
Werkzeug==0.14.1
wrapt==1.11.1
yagmail==0.11.214
zipp==0.3.3
